process_tilt_package/requirements.txt
build==0.10.0
click==8.1.3
Deprecated==1.2.13
docutils==0.19
Flask==2.2.2
fluent-logger==0.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
lorem-text==2.1
MarkupSafe==2.1.2
morecontext==0.6.0
msgpack==1.0.4
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-aws-lambda==0.37b0
opentelemetry-instrumentation-dbapi==0.37b0
opentelemetry-instrumentation-flask==0.37b0
opentelemetry-instrumentation-jinja2==0.37b0
opentelemetry-instrumentation-logging==0.37b0
opentelemetry-instrumentation-sqlite3==0.37b0
opentelemetry-instrumentation-urllib==0.37b0
opentelemetry-instrumentation-wsgi==0.37b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opentelemetry-util-http==0.37b0
packaging==23.0
pyproject_hooks==1.0.0
tilt==0.0.5
tomli==2.0.1
typing_extensions==4.5.0
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.15.0
